@@ -20,81 +20,86 @@
   },
   "resolutions": {
     "file-saver": "2.0.5",
-    "rxjs": "7.5.5",
-    "nx": "^14"
+    "rxjs": "7.5.7"
   },
   "dependencies": {
-    "@docusaurus/core": "2.0.0-beta.13",
-    "@docusaurus/preset-classic": "2.0.0-beta.13",
-    "@mdx-js/react": "^1.6.21",
-    "clsx": "^1.1.1",
-    "concurrently": "^7.1.0",
+    "@angular-devkit/core": "^14.2.6",
+    "@angular/core": "~14.2.6",
+    "@docusaurus/core": "2.1.0",
+    "@docusaurus/preset-classic": "2.1.0",
+    "@mdx-js/react": "^2.1.5",
+    "@storybook/builder-webpack5": "~6.5.12",
+    "@storybook/core-server": "~6.5.12",
+    "@storybook/manager-webpack5": "~6.5.12",
+    "clsx": "^1.2.1",
+    "concurrently": "^7.4.0",
     "env-cmd": "^10.1.0",
-    "prism-react-renderer": "^1.2.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "tslib": "^2.3.0"
+    "prism-react-renderer": "^1.3.5",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "tslib": "^2.4.0"
   },
   "devDependencies": {
-    "@angular-architects/module-federation": "^14.2.3",
-    "@angular-devkit/schematics": "^13.3.4",
-    "@commitlint/cli": "^16.2.4",
-    "@commitlint/config-angular": "^16.2.4",
-    "@commitlint/config-conventional": "^16.2.4",
-    "@docusaurus/module-type-aliases": "2.0.0-beta.13",
-    "@jscutlery/semver": "^2.23.3",
-    "@nrwl/angular": "^14.0.5",
-    "@nrwl/cli": "14.0.5",
-    "@nrwl/cypress": "^14.0.5",
-    "@nrwl/devkit": "^14.1.5",
-    "@nrwl/eslint-plugin-nx": "^14.0.5",
-    "@nrwl/jest": "^14.0.5",
-    "@nrwl/js": "^14.0.5",
-    "@nrwl/linter": "^14.0.5",
-    "@nrwl/nest": "^14.0.5",
-    "@nrwl/node": "^14.0.5",
-    "@nrwl/nx-cloud": "latest",
-    "@nrwl/nx-plugin": "^14.0.5",
-    "@nrwl/react": "^14.0.5",
-    "@nrwl/storybook": "^14.0.5",
-    "@nrwl/tao": "^14.0.5",
-    "@nrwl/web": "^14.0.5",
-    "@nrwl/workspace": "^14.0.5",
+    "@angular-architects/module-federation": "^14.3.12",
+    "@angular-devkit/schematics": "14.2.6",
+    "@commitlint/cli": "^17.1.2",
+    "@commitlint/config-angular": "^17.1.0",
+    "@commitlint/config-conventional": "^17.1.0",
+    "@docusaurus/module-type-aliases": "2.1.0",
+    "@jscutlery/semver": "^2.27.1",
+    "@nrwl/angular": "15.0.0",
+    "@nrwl/cli": "15.0.0",
+    "@nrwl/cypress": "15.0.0",
+    "@nrwl/devkit": "15.0.0",
+    "@nrwl/eslint-plugin-nx": "15.0.0",
+    "@nrwl/jest": "15.0.0",
+    "@nrwl/js": "15.0.0",
+    "@nrwl/linter": "15.0.0",
+    "@nrwl/nest": "15.0.0",
+    "@nrwl/node": "15.0.0",
+    "@nrwl/nx-cloud": "14.7.0",
+    "@nrwl/nx-plugin": "15.0.0",
+    "@nrwl/react": "15.0.0",
+    "@nrwl/storybook": "15.0.0",
+    "@nrwl/web": "15.0.0",
+    "@nrwl/workspace": "15.0.0",
     "@nx-clean/plugin-core": "^1.7.0",
-    "@nx-dotnet/core": "^1.9.12",
-    "@nx-go/nx-go": "^2.5.0",
-    "@nx-plus/docusaurus": "^13.0.1",
-    "@nx-tools/nx-docker": "^2.3.0",
-    "@nxrs/cargo": "^0.3.2",
-    "@twittwer/compodoc": "^1.7.0",
-    "@types/jest": "27.4.1",
-    "@types/node": "14.14.33",
+    "@nx-dotnet/core": "^1.14.0",
+    "@nx-go/nx-go": "^2.7.0",
+    "@nx-plus/docusaurus": "^14.1.0",
+    "@nx-tools/nx-docker": "^3.0.4",
+    "@nxrs/cargo": "^0.3.3",
+    "@swc-node/register": "^1.4.2",
+    "@swc/core": "^1.2.173",
+    "@twittwer/compodoc": "^1.8.2",
+    "@types/jest": "28.1.8",
+    "@types/node": "18.11.0",
     "@types/shelljs": "^0.8.11",
-    "@typescript-eslint/eslint-plugin": "~5.18.0",
-    "@typescript-eslint/parser": "~5.18.0",
-    "@yarnpkg/plugin-dlx": "^4.0.0-rc.4",
-    "all-contributors-cli": "^6.20.0",
-    "dotenv": "^16.0.0",
-    "esbuild": "^0.14.38",
-    "eslint": "~8.12.0",
+    "@typescript-eslint/eslint-plugin": "5.40.1",
+    "@typescript-eslint/parser": "5.40.1",
+    "all-contributors-cli": "^6.24.0",
+    "dotenv": "^16.0.3",
+    "esbuild": "^0.15.11",
+    "eslint": "8.25.0",
     "eslint-config-prettier": "^8.5.0",
-    "husky": "^7.0.4",
-    "jest": "27.5.1",
-    "jest-extended": "^2.0.0",
+    "husky": "^8.0.1",
+    "jest": "28.1.3",
+    "jest-extended": "^3.1.0",
+    "nx": "15.0.0",
     "nx-plugin-esbuild": "^0.5.0",
-    "pm2": "^5.2.0",
-    "prettier": "^2.3.1",
+    "pm2": "^5.2.2",
+    "prettier": "2.7.1",
     "shelljs": "^0.8.5",
-    "ts-jest": "27.1.4",
-    "ts-node": "9.1.1",
-    "tslib": "^2.0.0",
-    "typescript": "~4.3.5",
-    "webpack": "^5.72.0"
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
+    "tslib": "^2.4.0",
+    "typescript": "4.8.4",
+    "webpack": "^5.74.0"
   },
-  "packageManager": "yarn@3.2.0",
+  "packageManager": "pnpm@7.13.0",
   "engines": {
-    "node": ">=16.0.0",
+    "node": ">=18.0.0",
     "npm": ">=8.0.0",
-    "yarn": ">=1.22.0"
+    "pnpm": ">=7.13.0"
   }
 }
